requirements.txt
alembic>=1.13
fastapi>=0.110
jinja2>=3.1
passlib[bcrypt]>=1.7
psycopg2-binary>=2.9
pydantic-settings>=2.2
python-multipart>=0.0.9
sqlalchemy>=2.0
uvicorn[standard]>=0.27
